pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "QuickUp"
dynamic = ["version"]
requires-python = ">=3.10"
description = " A colorful command-line client for Clickup written in Python 3.10+."
readme = "README.md"
authors = [
    { name = "Matias Agustin Mendez", email = "matagus@gmail.com" },
]
keywords = [
  "clickup", "console", "python", "cli", "command-line", "task-list", "task-manager", "task-list-manager",
  "terminal-based", "clickup-api", "console-app", "inquirer", "clickup-client", "clickup-cli",
]

classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: English",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Software Development",
]
dependencies = [
  "python-dotenv", "colorist", "pyclickup", "inquirer", "cyclopts"
]

[project.optional-dependencies]
dev = [
  "pytest", "pytest-mock", "pytest-cov", "coverage"
]
docs = [
  "sphinx>=7.1.2,<8",
  "sphinx_rtd_theme>=2.0.0",
  "gitpython>=3.1.31",
  "sphinx-copybutton>=0.5.2",
]

[project.scripts]
quickup = "quickup.cli.main:run_app"

[project.urls]
Homepage = "https://github.com/matagus/quickup"
Documentation = "https://quickup.readthedocs.io/en/latest/"
Issues = "https://github.com/matagus/quickup/issues"
Changelog = "https://github.com/matagus/quickup/releases"
Repository = "https://github.com/matagus/quickup"

[tool.hatch.version]
path = "__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "ipython", "ipdb"
]

[tool.hatch.envs.test]
dependencies = [
    "pytest", "pytest-mock", "pytest-cov"
]

[tool.hatch.envs.default.scripts]
app = "python quickup.py"
install = "pip install -e .[dev]"

[tool.hatch.envs.docs]
dependencies = [
    "sphinx>=7.1.2,<8",
    "sphinx_rtd_theme>=2.0.0",
    "gitpython>=3.1.31",
    "sphinx-copybutton>=0.5.2",
]

[tool.hatch.envs.docs.scripts]
build = "sphinx-build -b html docs/source docs/build"
serve = "python -m http.server 8000 --directory docs/build"

[tool.hatch.envs.debug]
dependencies = ["ipdb>=0.13.9", "line_profiler>=3.5.1"]

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12", "3.13", "3.14"]

[tool.hatch.envs.test.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = ["coverage json", "coverage report"]
cov = ["test-cov", "cov-report"]
app = "python quickup.py"

[tool.ruff]
line-length = 120

[tool.ruff.lint]
fixable = [
    "ALL",
]
unfixable = []
select = ["E", "F", "UP", "B", "SIM", "PLC0415"]

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120
known_first_party = ["quickup"]
src_paths = ["."]
force_sort_within_sections = true

[tool.pyright]
include = ["quickup", "docs/source/conf.py"]
exclude = ["**/__pycache__"]

[tool.pytest.ini_options]
addopts = "-v --tb=short"
testpaths = ["tests"]

[tool.coverage.run]
branch = true
source = ["quickup"]
omit = ["*/tests/*", "*/__pycache__/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
]
